Almost every NPC ingame is now aware of Faulty Firmware Upgrade.
This doesn't means they'll all give you congratulations - some of them
have their own issues to mind, like preventing bots which are not from
HF from teletransporting in, etc.

Bob won't say about being murdered while sleeping, Bruce will dismiss you
because he finally can rest, Chandra won't claim there are active bots but
say there used to be, Dixon will speak about how much he wants to travel but
he will still kill you if you're on his black list, Erin will say she might
have seen a 123 bot, Jennifer will be too busy trying to make bots work again
because boss might return, Karol will notice diminshed commerce, Lukas will
try to convice you to buy things but won't sell you an Exterminator, Peter
will say nonsense about a civilian war, Ms Stone will give you a heads up,
GuardHouse guard will say about bad omen, North gate guard will stay in guard,
because it was at times of peace when his sister died, and WillGapes will,
well, be paranoid about going to town.

That's almost everyone accounted for, with very few exceptions like John and Ewald,
which does not have a real reason to react about it.
A tabulation error on Francis was also fixed.

+++ data/storyline/act1/dialogs/Bob.lua
@@ -137,7 +137,11 @@ return {
 			Tux:says(_"Here you go.")
 			Tux:del_item("Red Dilithium Crystal", 1)
 			Npc:says(_"Oh dude, you can't even imagine how thankful I am!")
-			Npc:says(_"My bot is now fine: the power level is holding steady at 9001 megawatts. I won't need to worry that bots will kill me while I am sleeping! Thank you so much!")
+			if (not HF_FirmwareUpdateServer_uploaded_faulty_firmware_update) then
+				Npc:says(_"My bot is now fine: the power level is holding steady at 9001 megawatts. I won't need to worry that bots will kill me while I am sleeping! Thank you so much!")
+			else
+				Npc:says(_"My bot is now fine: the power level is holding steady at 9001 megawatts. Now I have a robot and my brother to talk to, even knowing none of them will reply my questions! Thank you so much!")
+			end
 			Npc:says(_"Here is my award for you. When we saw this place at the first time, we found some bot inside. I destroyed it and this mace dropped out of it. I'm not sure how did it appear inside, but now it's yours.")
 			Tux:add_item("Mace")
 			Tux:says(_"Thank you, Bob.")
